u500vc707devkit: reconnect JTAG
If JTAG is not connected, sys_reset = reset.
This is a problem, because sys_reset is given to both PLLs and the MIG.
The MIG has an internal IBUF which then breaks everything.

Leaving JTAG on makes sys_reset = reset | jtag_reset
As this is not a top-level pin, the IBUF is ignored and it works.